+++ include/zephyr/drivers/rtc.h
@@ -197,13 +197,6 @@ __subsystem struct rtc_driver_api {
  */
 __syscall int rtc_set_time(const struct device *dev, const struct rtc_time *timeptr);
 
-static inline int z_impl_rtc_set_time(const struct device *dev, const struct rtc_time *timeptr)
-{
-	const struct rtc_driver_api *api = (const struct rtc_driver_api *)dev->api;
-
-	return api->set_time(dev, timeptr);
-}
-
 /**
  * @brief API for getting RTC time.
  *
@@ -215,13 +208,6 @@ static inline int z_impl_rtc_set_time(const struct device *dev, const struct rtc
  * @return -errno code if failure
  */
 __syscall int rtc_get_time(const struct device *dev, struct rtc_time *timeptr);
-
-static inline int z_impl_rtc_get_time(const struct device *dev, struct rtc_time *timeptr)
-{
-	const struct rtc_driver_api *api = (const struct rtc_driver_api *)dev->api;
-
-	return api->get_time(dev, timeptr);
-}
 
 /**
  * @name RTC Interface Alarm
@@ -245,18 +231,6 @@ static inline int z_impl_rtc_get_time(const struct device *dev, struct rtc_time 
  */
 __syscall int rtc_alarm_get_supported_fields(const struct device *dev, uint16_t id,
 					     uint16_t *mask);
-
-static inline int z_impl_rtc_alarm_get_supported_fields(const struct device *dev, uint16_t id,
-							uint16_t *mask)
-{
-	const struct rtc_driver_api *api = (const struct rtc_driver_api *)dev->api;
-
-	if (api->alarm_get_supported_fields == NULL) {
-		return -ENOSYS;
-	}
-
-	return api->alarm_get_supported_fields(dev, id, mask);
-}
 
 /**
  * @brief API for setting RTC alarm time.
@@ -284,18 +258,6 @@ static inline int z_impl_rtc_alarm_get_supported_fields(const struct device *dev
 __syscall int rtc_alarm_set_time(const struct device *dev, uint16_t id, uint16_t mask,
 				 const struct rtc_time *timeptr);
 
-static inline int z_impl_rtc_alarm_set_time(const struct device *dev, uint16_t id, uint16_t mask,
-					    const struct rtc_time *timeptr)
-{
-	const struct rtc_driver_api *api = (const struct rtc_driver_api *)dev->api;
-
-	if (api->alarm_set_time == NULL) {
-		return -ENOSYS;
-	}
-
-	return api->alarm_set_time(dev, id, mask, timeptr);
-}
-
 /**
  * @brief API for getting RTC alarm time.
  *
@@ -314,18 +276,6 @@ static inline int z_impl_rtc_alarm_set_time(const struct device *dev, uint16_t i
 __syscall int rtc_alarm_get_time(const struct device *dev, uint16_t id, uint16_t *mask,
 				 struct rtc_time *timeptr);
 
-static inline int z_impl_rtc_alarm_get_time(const struct device *dev, uint16_t id, uint16_t *mask,
-					    struct rtc_time *timeptr)
-{
-	const struct rtc_driver_api *api = (const struct rtc_driver_api *)dev->api;
-
-	if (api->alarm_get_time == NULL) {
-		return -ENOSYS;
-	}
-
-	return api->alarm_get_time(dev, id, mask, timeptr);
-}
-
 /**
  * @brief API for testing if RTC alarm is pending.
  *
@@ -342,17 +292,6 @@ static inline int z_impl_rtc_alarm_get_time(const struct device *dev, uint16_t i
  * @return -errno code if failure
  */
 __syscall int rtc_alarm_is_pending(const struct device *dev, uint16_t id);
-
-static inline int z_impl_rtc_alarm_is_pending(const struct device *dev, uint16_t id)
-{
-	const struct rtc_driver_api *api = (const struct rtc_driver_api *)dev->api;
-
-	if (api->alarm_is_pending == NULL) {
-		return -ENOSYS;
-	}
-
-	return api->alarm_is_pending(dev, id);
-}
 
 /**
  * @brief API for setting alarm callback.
@@ -382,18 +321,6 @@ static inline int z_impl_rtc_alarm_is_pending(const struct device *dev, uint16_t
  */
 __syscall int rtc_alarm_set_callback(const struct device *dev, uint16_t id,
 				     rtc_alarm_callback callback, void *user_data);
-
-static inline int z_impl_rtc_alarm_set_callback(const struct device *dev, uint16_t id,
-						rtc_alarm_callback callback, void *user_data)
-{
-	const struct rtc_driver_api *api = (const struct rtc_driver_api *)dev->api;
-
-	if (api->alarm_set_callback == NULL) {
-		return -ENOSYS;
-	}
-
-	return api->alarm_set_callback(dev, id, callback, user_data);
-}
 
 #endif /* CONFIG_RTC_ALARM */
 /**
@@ -428,18 +355,6 @@ static inline int z_impl_rtc_alarm_set_callback(const struct device *dev, uint16
 __syscall int rtc_update_set_callback(const struct device *dev, rtc_update_callback callback,
 				      void *user_data);
 
-static inline int z_impl_rtc_update_set_callback(const struct device *dev,
-						 rtc_update_callback callback, void *user_data)
-{
-	const struct rtc_driver_api *api = (const struct rtc_driver_api *)dev->api;
-
-	if (api->update_set_callback == NULL) {
-		return -ENOSYS;
-	}
-
-	return api->update_set_callback(dev, callback, user_data);
-}
-
 #endif /* CONFIG_RTC_UPDATE */
 /**
  * @}
@@ -471,17 +386,6 @@ static inline int z_impl_rtc_update_set_callback(const struct device *dev,
  */
 __syscall int rtc_set_calibration(const struct device *dev, int32_t calibration);
 
-static inline int z_impl_rtc_set_calibration(const struct device *dev, int32_t calibration)
-{
-	const struct rtc_driver_api *api = (const struct rtc_driver_api *)dev->api;
-
-	if (api->set_calibration == NULL) {
-		return -ENOSYS;
-	}
-
-	return api->set_calibration(dev, calibration);
-}
-
 /**
  * @brief API for getting RTC calibration.
  *
@@ -493,17 +397,6 @@ static inline int z_impl_rtc_set_calibration(const struct device *dev, int32_t c
  * @return -errno code if failure
  */
 __syscall int rtc_get_calibration(const struct device *dev, int32_t *calibration);
-
-static inline int z_impl_rtc_get_calibration(const struct device *dev, int32_t *calibration)
-{
-	const struct rtc_driver_api *api = (const struct rtc_driver_api *)dev->api;
-
-	if (api->get_calibration == NULL) {
-		return -ENOSYS;
-	}
-
-	return api->get_calibration(dev, calibration);
-}
 
 #endif /* CONFIG_RTC_CALIBRATION */
 /**
@@ -555,6 +448,7 @@ static inline int32_t rtc_calibration_from_frequency(uint32_t frequency)
 }
 #endif
 
+#include <zephyr/drivers/rtc/internal/rtc_impl.h>
 #include <zephyr/syscalls/rtc.h>
 
 #endif /* ZEPHYR_INCLUDE_DRIVERS_RTC_H_ */
